+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
+## [0.2.1] - 2020-05-10
+### Changed
+- Lock cuttlepool version at 0.6.0.
+
 ## [0.2.0] - 2018-01-30
 ### Added
 - Import `PoolConnection` for simple, uniform import from `flask_cuttlepool`.
@@ -23,4 +27,6 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ### Added
 - Initial Flask-CuttlePool code.
 
+[Unreleased]: https://github.com/smitchell556/flask-cuttlepool/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/smitchell556/flask-cuttlepool/compare/v0.2.0...0.2.1
 [0.2.0]: https://github.com/smitchell556/flask-cuttlepool/compare/v0.1.0...0.2.0